requirements.txt
anyio==4.13.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
dash==4.1.0
dash-bootstrap-components==2.0.4
Flask==3.1.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==9.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.19.0
nest-asyncio==1.6.0
numpy==2.4.4
packaging==26.0
pandas==3.0.2
plotly==6.7.0
plotly-cloud==0.2.1
psycopg2-binary==2.9.11
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
retrying==1.4.2
rich==15.0.0
setuptools==82.0.1
six==1.17.0
SQLAlchemy==2.0.49
tomli==2.4.1
tomli_w==1.2.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
zipp==3.23.0
